refactor: move general to hist
